Updated the PADTextureExtractor to handle compressed textures
Starting with version 7.8.1, the Puzzle & Dragons servers began serving
monster textures in a compressed format. While this was good for the
memory footprint of the game, it also meant that any textures downloaded
within the last six months could not be extracted by the
PADTextureExtractor.

This update adds the necessary functionality to handle compressed
textures. Happy extracting!